Fix repository name
